+++ src/wifi/model/he/he-phy.h
@@ -223,8 +223,8 @@ class HePhy : public VhtPhy
      * \param staId the STA-ID of the station taking part of the UL MU
      * \return the center frequency in MHz corresponding to the non-HE portion of the HE TB PPDU
      */
-    std::vector<uint16_t> GetCenterFrequenciesForNonHePart(const Ptr<const WifiPpdu> ppdu,
-                                                           uint16_t staId) const;
+    std::vector<int64_t> GetCenterFrequenciesForNonHePart(const Ptr<const WifiPpdu> ppdu,
+                                                          uint16_t staId) const;
 
     /**
      * Sets the OBSS-PD algorithm.
@@ -459,7 +459,7 @@ class HePhy : public VhtPhy
     static std::vector<WifiSpectrumBandIndices> ConvertHeRuSubcarriers(
         ChannelWidthMhz bandWidth,
         ChannelWidthMhz guardBandwidth,
-        const std::vector<uint16_t>& centerFrequencies,
+        const std::vector<int64_t>& centerFrequencies,
         ChannelWidthMhz totalWidth,
         uint32_t subcarrierSpacing,
         HeRu::SubcarrierRange subcarrierRange,
