+++ src/wifi/test/spectrum-wifi-phy-test.cc
@@ -669,7 +669,7 @@ SpectrumWifiPhyFilterTest::DoTeardown()
 void
 SpectrumWifiPhyFilterTest::RunOne()
 {
-    uint16_t txFrequency;
+    int64_t txFrequency;
     switch (m_txChannelWidth)
     {
     case 20:
@@ -695,7 +695,7 @@ SpectrumWifiPhyFilterTest::RunOne()
     m_txPhy->SetOperatingChannel(
         WifiPhy::ChannelTuple{txChannelNum, m_txChannelWidth, WIFI_PHY_BAND_5GHZ, 0});
 
-    uint16_t rxFrequency;
+    int64_t rxFrequency;
     switch (m_rxChannelWidth)
     {
     case 20:
@@ -968,8 +968,8 @@ SpectrumWifiPhyGetBandTest::DoRun()
                 expectedStartIndice + (indicesPer20MhzBand * (bandWidth / 20)) - 1;
             std::vector<WifiSpectrumBandIndices> expectedIndices{
                 {expectedStartIndice, expectedStopIndice}};
-            const uint64_t expectedStartFrequency = 5170 * 1e6;
-            const uint64_t expectedStopFrequency = (5170 + bandWidth) * 1e6;
+            const int64_t expectedStartFrequency = 5170 * 1e6;
+            const int64_t expectedStopFrequency = (5170 + bandWidth) * 1e6;
             std::vector<WifiSpectrumBandFrequencies> expectedFrequencies{
                 {expectedStartFrequency, expectedStopFrequency}};
             const std::size_t numBands = (channelWidth / bandWidth);
@@ -1289,7 +1289,7 @@ class SpectrumWifiPhy80Plus80Test : public TestCase
      * \param expectSuccess flag to indicate whether reception is expected to be successful
      */
     void RunOne(const std::vector<uint8_t>& channelNumbers,
-                uint16_t interferenceCenterFrequency,
+                int64_t interferenceCenterFrequency,
                 ChannelWidthMhz interferenceBandWidth,
                 bool expectSuccess);
 
@@ -1512,7 +1512,7 @@ SpectrumWifiPhy80Plus80Test::DoTeardown()
 
 void
 SpectrumWifiPhy80Plus80Test::RunOne(const std::vector<uint8_t>& channelNumbers,
-                                    uint16_t interferenceCenterFrequency,
+                                    int64_t interferenceCenterFrequency,
                                     ChannelWidthMhz interferenceBandWidth,
                                     bool expectSuccess)
 {
